(chore) deprecate C-like
